refactor gate for multi-guild, add base of channel directory plugin

+++ basement_bot/plugins/gate.py
@@ -14,12 +14,12 @@ class ServerGate(cogs.MatchPlugin):
     PLUGIN_NAME = __name__
 
     async def preconfig(self):
-        self.channel = self.bot.get_channel(self.config.channel)
-        if not self.channel:
-            raise RuntimeError("Could not find channel to use as server gate")
+        self.channels = set()
+        for channel_config in self.config.values():
+            self.channels.add(channel_config.channel)
 
     async def match(self, ctx, _):
-        if ctx.channel.id != self.channel.id:
+        if not ctx.channel.id in self.channels:
             return False
         return True
 
@@ -27,32 +27,41 @@ class ServerGate(cogs.MatchPlugin):
         if content.startswith(self.bot.command_prefix):
             return
 
-        bot_message = None
+        guild_config = self.config.get(ctx.guild.id)
+        if not guild_config:
+            return
 
-        if content.lower() == "agree":
+        await ctx.message.delete()
+
+        if content.lower() == guild_config.get("passing_text", "agree"):
             roles = await self.get_roles(ctx)
             if not roles:
                 return
 
             await ctx.author.add_roles(*roles)
 
-            bot_message = await self.bot.h.tagged_response(
+            welcome_message = guild_config.get(
+                "welcome_message", "Welcome to the server!"
+            )
+            delete_wait = guild_config.get("delete_wait_seconds", 30)
+
+            bot_message = await self.tagged_response(
                 ctx,
-                f"{self.config.welcome_message} (this message will delete in {self.config.delete_wait_seconds} seconds)",
+                f"{welcome_message} (this message will delete in {delete_wait} seconds)",
             )
 
-        await ctx.message.delete()
-
-        if not bot_message:
-            return
-
-        await asyncio.sleep(self.config.delete_wait_seconds)
-        await bot_message.delete()
+            await asyncio.sleep(delete_wait)
+            await bot_message.delete()
 
     async def get_roles(self, ctx):
         roles = []
-        for role_name in self.config.roles:
+        config_roles = self.config.get(ctx.guild.id, {}).get("roles", [])
+        for role_name in config_roles:
             role = discord.utils.get(ctx.guild.roles, name=role_name)
+
+            if role in ctx.author.roles:
+                continue
+
             if role:
                 roles.append(role)
